+++ test/Microsoft.ApplicationInsights.AspNetCore.Tests/Helpers/CommonMocks.cs
@@ -1,7 +1,6 @@
 ﻿namespace Microsoft.ApplicationInsights.AspNetCore.Tests.Helpers
 {
     using System;
-    using DiagnosticListeners;
     using Microsoft.ApplicationInsights.Channel;
     using Microsoft.ApplicationInsights.Extensibility;
 
@@ -9,6 +8,7 @@
     {
         public const string InstrumentationKey = "REQUIRED";
         public const string InstrumentationKeyHash = "0KNjBVW77H/AWpjTEcI7AP0atNgpasSkEll22AtqaVk=";
+        public const string TestApplicationId = nameof(TestApplicationId);
 
         public static TelemetryClient MockTelemetryClient(Action<ITelemetry> onSendCallback)
         {
@@ -19,9 +19,9 @@
             });
         }
 
-        internal static ICorrelationIdLookupHelper MockCorrelationIdLookupHelper()
+        internal static IApplicationIdProvider GetMockApplicationIdProvider()
         {
-            return new CorrelationIdLookupHelperStub();
+            return new MockApplicationIdProvider(InstrumentationKey, TestApplicationId);
         }
     }
 }